requirements.txt
psycopg2==2.7.1
peewee==2.9.1
requests==2.13.0
pytz==2016.10
python-dateutil>=2.6.0
Scrapy==1.3.3
xlrd>=1.0.0
schedule>=0.4.2